Rewrite PictureSet with TURBO!
This changes how partial invals are done by adding a hybrid mode.

What we used to do is generate a SkPicture for the new area. This
SkPicture would possibly be larger than the actual inval, depending
on various merge rules (more SkPictures == slower to draw a tile)

The new code rewrites PictureSet entirely, preserving many of the old rules
but cleans up the code and adds the concept of a "PrerenderedInval". This is
a partial inval that WebKit has rasterized. By having WebKit produce both
a SkPicture and a SkBitmap, we avoid needing to play back the picture and
avoid overdrawing. We take this SkBitmap, and simply update the front
textures with it. This gives us full partial invals through the entire
system without hitting any driver bugs, and with minimal copies. And while
the SkPicture may be larger than the inval, the SkBitmap that is rasterized
is not - it matches the area webkit has said is dirty.

+++ Source/WebCore/platform/graphics/android/rendering/TilesManager.cpp
@@ -471,12 +471,12 @@ bool TilesManager::updateContextIfChanged()
     return changed;
 }
 
-float TilesManager::tileWidth()
+int TilesManager::tileWidth()
 {
     return TILE_WIDTH;
 }
 
-float TilesManager::tileHeight()
+int TilesManager::tileHeight()
 {
     return TILE_HEIGHT;
 }
